pyproject.toml
[project]
name = "savings-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=4.0.0",
    "alembic>=1.15.1",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "firebase-admin>=6.7.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "paystackapi>=2.1.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "ruff>=0.11.2",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
